ポジションゲージの最初のラウンドの位置修正 ラウンド変更時に優勢ゲージ量を変更

+++ src/IntoTheAbyss/BossPatternAttack.h
@@ -18,6 +18,7 @@ private:
 	bool changeAngleFlag;
 	const float SHOT_BULLET_MAX;//ˆê“x‚ÉŒ‚‚Â’e‚Ì‘”
 	const float MAX_SPEED;//ˆê“x‚ÉŒ‚‚Â’e‚Ì‘”
-	const float SHOT_INTERVAL_MAX;//’e‚ğŒ‚‚ÂŠÔŠu
+	const int SHOT_INTERVAL_MAX;//’e‚ğŒ‚‚ÂŠÔŠu
+
 };
